requirements.txt
fastapi
uvicorn
pydantic
email-validator
jinja2
python-multipart
fastapi-mail
python-dotenv
sqlalchemy
stripe
httpx
